LayoutTests/http/tests/device-orientation/device-motion-third-party-iframe-denied-expected.txt

CONSOLE MESSAGE: No device motion events will be fired, reason: Third-party iframes are not allowed access to device motion unless explicitly allowed via Feature-Policy (gyroscope & accelerometer).
CONSOLE MESSAGE: Call to requestPermission() failed, reason: Third-party iframes are not allowed access to device motion unless explicitly allowed via Feature-Policy (gyroscope & accelerometer).
Tests that third-party iframes cannot request access to device motion events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.data is "denied"
PASS successfullyParsed is true

TEST COMPLETE
